feat: redirect signed-in users from login to dashboard
